http://jira.aplana.com/browse/SBRFACCTAX-1862 сейчас можно видеть таблицу с евентами для формы. Осталось:1) таблицу привести к кастомному виду 2) добавить время а не только дату 3) добавить функционал для декларации 4) перенести добавление истории в серисы 5) связать создание истории с пользователем.
git-svn-id: https://horse.aplana.com/svn/SBRFACCTAX/trunk/taxaccounting@2121 34aa7802-8fb0-764d-beb8-9b5b9d595954

+++ core/src/main/java/com/aplana/sbrf/taxaccounting/service/LogBusinessService.java
@@ -13,14 +13,14 @@ public interface LogBusinessService {
 	 * @param declarationId идентификатор декларации
 	 * @return объект, представляющий историю событий для декларации
 	 */
-	List<LogBusiness> getDeclarationLogsBusiness(int declarationId);
+	List<LogBusiness> getDeclarationLogsBusiness(long declarationId);
 
 	/**
 	 * Получить информацию об историей событий налоговой формы
 	 * @param formId идентификатор формы
 	 * @return объект, представляющий историю событий для налоговой формы
 	 */
-	List<LogBusiness> getFormLogsBusiness(int formId);
+	List<LogBusiness> getFormLogsBusiness(long formId);
 
 	/**
 	 * добавить информацию в историю событий декларации/налоговой формы
